+++ CHANGELOG.md
@@ -45,6 +45,8 @@ packages. Users of `selectrum-prescient` can update to configure
   deriving from `comint-mode` ([#459], [#462]).
 * `selectrum-completion-in-region` does support cycling (as configured
   per `completion-cycle-threshold`) now ([#419], [#456]).
+* The face that is used when the mouse hovers over candidates can now
+  be customized through `selectrum-mouse-highlight`.
 
 ### Bugs fixed
 * When completion functions would pass a list as default value,